ml-service/requirements.txt
# numpy/scipy first — scikit-surprise 1.1.x does not build against numpy 2.x
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.29.0
scikit-learn==1.4.2
scikit-surprise==1.1.3
sqlalchemy==2.0.30
pymysql==1.1.0
pydantic==2.7.1
apscheduler==3.10.4
joblib==1.4.2
